Created separate api files for spec_files and reference_files
